feat/GP-95WEB - Implement to view instructions information about the Writing test.
