+++ src/app/routes/PrivateRoute.jsx
@@ -2,7 +2,13 @@
 import HomePage from "@pages/HomePage.jsx";
 import IntroReading from "@pages/Reading/IntroductionScreen.jsx";
 import { ProtectedRoute } from "./ProtectedRoute/ProtectedRoute.jsx";
+import SpeakingPage from "@pages/SpeakingPage.jsx";
+import Introduction from "@features/speaking/ui/Introduction.jsx";
+import ReadingLayout from "@features/reading/ui/Layout.jsx";
+import WelcomeScreen from "@pages/WelcomeScreen.jsx";
+import Layout from "@pages/Layout.jsx";
 import WritingPage from "@pages/WritingPage.jsx";
+import IntroWriting from "@features/writing/ui/IntroWriting.jsx";
 
 const PrivateRoute = [
   {
@@ -30,6 +36,16 @@ const PrivateRoute = [
           {
             path: "intro",
             element: <IntroReading />,
+          },
+        ],
+      },
+      {
+        path: "writing",
+        element: <WritingPage />,
+        children: [
+          {
+            path: "intro",
+            element: <IntroWriting />,
           },
         ],
       },
